requirements.txt
fastapi~=0.82.0
holidays~=0.15
pyhumps~=3.7.3
pre-commit~=2.20.0
pycountry==~22.3.5
pydantic~=1.10.1
uvicorn~=0.18.3
